chore(docs): move Head to a new file
