+++ docs/src/pages/_app.page.tsx
@@ -1,19 +1,13 @@
 import * as React from 'react';
 
 import { ThemeProvider, ColorMode, defaultTheme } from '@aws-amplify/ui-react';
-import { configure, trackPageVisit } from '../utils/track';
 
-import Head from 'next/head';
+import { configure, trackPageVisit } from '@/utils/track';
 import { Header } from '@/components/Layout/Header';
 import Script from 'next/script';
 import { baseTheme } from '../theme';
-import { capitalizeString } from '@/utils/capitalizeString';
 import { useCustomRouter } from '@/components/useCustomRouter';
-import metaData from '@/data/pages.preval';
-import { PREVIEW_HEIGHT, PREVIEW_WIDTH } from '@/data/preview';
-import { FRAMEWORKS } from '@/data/frameworks';
-
-import { getImagePath } from '@/utils/previews';
+import { Head } from './Head';
 
 import '../styles/index.scss';
 
@@ -39,30 +33,10 @@ function MyApp({ Component, pageProps }) {
   const [expanded, setExpanded] = React.useState(false);
 
   const {
-    asPath,
     pathname,
     query: { platform = 'react' },
   } = useCustomRouter();
-  const asPathname = pathname.replace('[platform]', String(platform));
   const isHomepage = pathname === '/' || pathname === '/[platform]';
-  const filepath = `/${pathname
-    .split('/')
-    .filter((n) => n && n !== '[platform]')
-    .join('/')}`;
-  const { title, metaTitle, description, metaDescription } =
-    metaData[pathname]?.frontmatter ?? {};
-
-  if ((!description && !metaDescription) || (!title && !metaTitle)) {
-    throw new Error(`Meta Info missing on ${filepath}`);
-  }
-
-  const pageTitle = `${metaTitle ?? title} | ${capitalizeString(
-    platform
-  )} - Amplify UI`;
-  const homepagePaths = [
-    '/',
-    ...FRAMEWORKS.map((framework) => `/${framework}`),
-  ];
 
   const [colorMode, setColorMode] = React.useState<ColorMode>('system');
   const handleColorModeChange = (colorMode: ColorMode) => {
@@ -86,32 +60,8 @@ function MyApp({ Component, pageProps }) {
 
   return (
     <>
-      <Head>
-        <title>{pageTitle}</title>
-        {homepagePaths.includes(asPath) && (
-          <link rel="canonical" href="https://ui.docs.amplify.aws/" />
-        )}
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta name="description" content={metaDescription ?? description} />
-        <meta property="og:title" content={pageTitle} />
-        <meta property="og:type" content="object" />
-        <meta property="og:url" content={`${process.env.SITE_URL}${asPath}`} />
-        <meta
-          property="og:image"
-          content={process.env.SITE_URL + getImagePath(asPathname)}
-        />
-        <meta property="og:image:width" content={String(PREVIEW_WIDTH)} />
-        <meta property="og:image:height" content={String(PREVIEW_HEIGHT)} />
-        <meta
-          property="og:image:secure_url"
-          content={process.env.SITE_URL + getImagePath(asPathname)}
-        />
-        <meta property="og:image:type" content="image/png" />
-        <meta
-          property="og:image:alt"
-          content={metaDescription ?? description}
-        />
-      </Head>
+      <Head />
+
       <div className={isHomepage ? `docs-home` : ''}>
         <ThemeProvider theme={baseTheme} colorMode={colorMode}>
           <Header
